Extract reusable "FeedStoreSpecsProtocol" helper methods into a shared scope so it can be used by other "FeedStoreProtocol" implementation tests
